services/context_engineering/pyproject.toml
[project]
name = "context-engineering"
version = "0.1.0"
description = "Turing Agents - Context Engineering"
readme = "README.md"
requires-python = ">=3.11,<3.14"
dependencies = [
  "pydantic==2.8.2",
  "python-dotenv==1.0.1",
  "prometheus-client==0.20.0",
  "neo4j==5.25.0",
  "rank-bm25==0.2.2",
  "numpy==1.26.4",
  "google-generativeai>=0.7.2",
  "rouge-score==0.1.2",
  "nltk==3.9.1",
  "bert-score==0.3.13",
  "torch>=2.1.0; platform_system != 'Darwin' or platform_machine != 'arm64'",
  "deepeval>=0.20.0",
]

[project.optional-dependencies]
llm = [
  "openai>=1.40.0",
  "anthropic>=0.34.0",
]

[project.scripts]
ctx-v2 = "scripts.main:main"

[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["libs", "scripts"]

[tool.pytest.ini_options]
markers = [
  "llm_judge: tests that call an external LLM-as-judge via DeepEval",
]
